requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25

langchain-text-splitters==0.3.8
langsmith==0.3.42
orjson==3.10.18
packaging==24.2
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==9.1.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
zstandard==0.23.0
langchain-community==0.3.24
google-search-results
google-ai-generativelanguage==0.6.15
google-generativeai==0.8.5
langchain-google-genai==2.0.10
pytest                                     
qdrant-client==1.9.1                       
langgraph>=0.0.85     
langchain-openai>=0.1.9 
